+++ src/aimanager/HISTORY.rst
@@ -3,8 +3,13 @@
 Release History
 ===============
 
+1.5.2b2
++++++++
+* ``az aimanager namespace modeldeployment``: Accept ``--ns`` as an alias of
+  ``--namespace-name``.
+
 1.5.2b1
-++++++
++++++++
 * Refactor validation code to make the name validators consistent
 
 1.5.1